Address zoned stats review comments
Move the legacy Stat accumulator out of the zoned builder and into file stats, where it is still needed for footer statistics. Also derive the layout registry from VortexSession when loading layout flatbuffers.
