+++ vortex-layout/src/layouts/zoned/mod.rs
@@ -21,7 +21,6 @@ pub mod zone_map;
 use std::sync::Arc;
 
 pub(crate) use builder::AggregateStatsAccumulator;
-pub(crate) use builder::StatsAccumulator;
 pub(crate) use builder::aggregate_partials;
 use prost::Message;
 pub use schema::MAX_IS_TRUNCATED;
